local storage improve

+++ src/common/services/jwt.service.ts
@@ -13,38 +13,39 @@ import { jwtDecode } from "jwt-decode";
 export class JwtService {
     constructor(private requestService: RequestService,
         private router: Router,) {
-        this.localStorageCheck();
         this.refreshTokenInterval();
+        addEventListener('storage', async (event) => {
+            if (this.localStorageToken == true) {
+                localStorage.setItem('accessToken' ,(event.oldValue as string));
+                await this.requestService.roomRemoveUserAll(this.decode().username as string);
+                await this.requestService.localStorageStrike(event.oldValue as string);
+                await this.blackListToken();
+                localStorage.clear();
+                await Swal.fire({
+                    title: "session error",
+                    text: "unauthorized activity detected, please login again",
+                    icon: "error",
+                    background: "#101416",
+                    color: "white",
+                });
+                this.router.navigate(['login']);
+            }
+        });
     }
 
     refreshToken: string = '';
+    localStorageToken: boolean = false;
 
-    localStorageChange$: Subject<void> = new Subject<void>();//emits the event
-    tempAccessToken: string = localStorage.getItem('accessToken') as string;
-    localStorageToken: boolean = false;//to check or not to check localstorage
-    localStorageSubscribe: Subscription = this.localStorageChange$.subscribe(async () => {//handle the event
-        await this.requestService.roomRemoveUserAll(this.decode().username as string);
-        await this.requestService.localStorageStrike(this.tempAccessToken as string);
-        await this.blackListToken();
-        this.setLocalStorageToken(false);
-        localStorage.clear();
-        await Swal.fire({
-            title: "session error",
-            text: "unauthorized activity detected, please login again",
-            icon: "error",
-            background: "#101416",
-            color: "white",
-        });
-        this.router.navigate(['login']);
-    });
-
+    public setLocalStorageToken(token: boolean): void {
+        this.localStorageToken = token;
+    }
 
     public async verifyUrl(): Promise<boolean> {
         return (await axios.post(`${Constants.AUTH_SERVICE}/tokens/verify/url`, { token: localStorage.getItem('accessToken') })).data;
     }
 
     public async verifyManagerUrl(): Promise<boolean> {
-        return (await axios.post(`${Constants.AUTH_SERVICE}/tokens/verify/manager/url`, { token: localStorage.getItem('accessToken')})).data;
+        return (await axios.post(`${Constants.AUTH_SERVICE}/tokens/verify/manager/url`, { token: localStorage.getItem('accessToken') })).data;
     }
 
     public async verifyToken(): Promise<boolean> {
@@ -57,27 +58,6 @@ export class JwtService {
 
     public async verifyEditorToken(): Promise<boolean> {
         return (await axios.get(`${Constants.AUTH_SERVICE}/tokens/verify/editor`, { headers: { Authorization: `Bearer ${localStorage.getItem('accessToken')}` } })).data;
-    }
-
-
-    public setLocalStorageToken(token: boolean): void {
-        this.tempAccessToken = localStorage.getItem('accessToken') as string;
-        this.localStorageToken = token;
-    }
-
-    
-
-    public localStorageCheck(): void {
-        setInterval(() => {
-            if (localStorage.getItem('accessToken') as string != this.tempAccessToken) {
-                if (this.localStorageToken == true) {
-                    this.localStorageChange$.next();
-                }
-                else {
-                    this.tempAccessToken = localStorage.getItem('accessToken') as string;
-                }
-            }
-        }, 3000);
     }
 
     public decode(): CustomJwtPayload {
@@ -99,16 +79,14 @@ export class JwtService {
     }
 
     public async getRefreshToken(): Promise<string> {
-        return (await axios.post(`${Constants.AUTH_SERVICE}/tokens/refresh/get`, {token: localStorage.getItem('accessToken')}, { headers: { Authorization: `Bearer ${localStorage.getItem('accessToken')}` } })).data;
+        return (await axios.post(`${Constants.AUTH_SERVICE}/tokens/refresh/get`, { token: localStorage.getItem('accessToken') }, { headers: { Authorization: `Bearer ${localStorage.getItem('accessToken')}` } })).data;
     }
 
     public refreshTokenInterval(): void {
         try {
             setInterval(async () => {
                 if (this.refreshToken != '') {
-                    this.setLocalStorageToken(false);
                     await this.refreshAccessToken();
-                    this.setLocalStorageToken(true);
                 }
             }, 288000);
         }
